Change from double to float for a modest speed boost.

+++ src/camera.h
@@ -3,13 +3,13 @@
 #include "vec3.h"
 
 struct Camera {
-    Camera(const point3& origin, const double& vpw, const double& vph, const double& focal_length):
+    Camera(const point3& origin, const float& vpw, const float& vph, const float& focal_length):
         origin(origin), vpw(vpw), vph(vph), focal_length(focal_length) {}
 
     point3 origin;
 
-    double vpw;
-    double vph;
-    double focal_length = 1.0;
+    float vpw;
+    float vph;
+    float focal_length = 1.0;
 
 };
