+++ src/rectangle.h
@@ -9,20 +9,20 @@
 enum face { x_side, y_side, z_side };
 
 struct Rectangle : Object {
-    Rectangle(const point3& o, const double& w, const double& h, face face, const Props& props):
+    Rectangle(const point3& o, const float& w, const float& h, face face, const Props& props):
                     Object(props), orig(o), w(w), h(h), face(face) {}
 
-    double hit(const ray& r) const;
+    float hit(const ray& r) const;
     vec3 normal(const point3& hp) const;
     inline bool check_bounds(const point3& hp) const;
     inline bool same_plane(const point3& hp) const;
 
     point3 orig;
-    double w, h;
+    float w, h;
     face face;
 };
 
-inline bool is_between(double x, double a, double b) {
+inline bool is_between(float x, float a, float b) {
     if (a > b) std::swap(a, b);
     return x >= a && x <= b;
 }
@@ -44,12 +44,12 @@ inline bool Rectangle::same_plane(const point3& hp) const {
 }
 
 
-inline double Rectangle::hit(const ray& r) const {
+inline float Rectangle::hit(const ray& r) const {
     if (face == x_side && r.dir.x() == 0.) return -1;
     if (face == y_side && r.dir.y() == 0.) return -1;
     if (face == z_side && r.dir.z() == 0.) return -1;
 
-    double t;
+    float t;
     if (face == x_side) t = (orig.x() - r.orig.x())/r.dir.x();
     else if (face == y_side) t = (orig.y() - r.orig.y())/r.dir.y();
     else t = (orig.z() - r.orig.z())/r.dir.z();
