+++ src/sphere.h
@@ -5,14 +5,14 @@
 #include "ray.h"
 
 struct Sphere : Object {
-    Sphere(const point3& center, const double& radius, const Props& props):
+    Sphere(const point3& center, const float& radius, const Props& props):
             Object(props), center(center), radius(radius) {}
 
-    double hit(const ray& r) const;
+    float hit(const ray& r) const;
     vec3 normal(const point3& hp) const;
 
     point3 center;
-    double radius;
+    float radius;
 };
 
 
@@ -20,13 +20,13 @@ struct Sphere : Object {
 // (R-C)(R-C) = normal(const point3& hp) constA + Bx) + C^2 - r^2 = 0
 // (Bx)^2 + (2AB - 2BC)x + A^2 + C^2 - r^2 - 2AC = 0
 // (Bx)^2 + 2B(A-C)x + (A-C)^2 - r^2 = 0
-inline double Sphere::hit(const ray& r) const {
+inline float Sphere::hit(const ray& r) const {
     vec3 oc = r.orig - this->center;
-    double a = r.dir * r.dir;
-    double b = 2 * r.dir * oc;
-    double c = oc*oc - this->radius * this->radius;
+    float a = r.dir * r.dir;
+    float b = 2 * r.dir * oc;
+    float c = oc*oc - this->radius * this->radius;
 
-    double disc = (b*b - 4*a*c);
+    float disc = (b*b - 4*a*c);
 
     if (disc < 0) return -1;
     return (-b - sqrt(disc))/(2*a);
